+++ tests/pedal/test_pedal.py
@@ -2,46 +2,9 @@ import time
 import unittest
 from panda import Panda
 from panda_jungle import PandaJungle  # pylint: disable=import-error
-
-
-class CanHandle(object):
-  def __init__(self, p):
-    self.p = p
-
-  def transact(self, dat):
-    self.p.isotp_send(1, dat, 0, recvaddr=2)
-
-    def _handle_timeout(signum, frame):
-      # will happen on reset
-      raise Exception("timeout")
-
-    signal.signal(signal.SIGALRM, _handle_timeout)
-    signal.alarm(1)
-    try:
-      ret = self.p.isotp_recv(2, 0, sendaddr=1)
-    finally:
-      signal.alarm(0)
-
-    return ret
-
-  def controlWrite(self, request_type, request, value, index, data, timeout=0):
-    # ignore data in reply, panda doesn't use it
-    return self.controlRead(request_type, request, value, index, 0, timeout)
-
-  def controlRead(self, request_type, request, value, index, length, timeout=0):
-    dat = struct.pack("HHBBHHH", 0, 0, request_type, request, value, index, length)
-    return self.transact(dat)
-
-  def bulkWrite(self, endpoint, data, timeout=0):
-    if len(data) > 0x10:
-      raise ValueError("Data must not be longer than 0x10")
-    dat = struct.pack("HH", endpoint, len(data)) + data
-    return self.transact(dat)
-
-  def bulkRead(self, endpoint, length, timeout=0):
-    dat = struct.pack("HH", endpoint, 0)
-    return self.transact(dat)
-    
+import os
+import sys
+from can_helper import CanHandle
 
 class TestPedal(unittest.TestCase):
   PEDAL_BUS = 1
